Refactored model loading.
Model is now loaded into vectors on the stack before stuffing into
openGL. ModelGL is now a much lighter class, and knows much less of GL's
behind the scenes workings.

Fixed up codes const behaviour in a few places, to prevent confusion.

Folded the loading loop in together as iterating the file twice is a
waste if it doesn't need to be done.

fixed a minor bug that would have model files with any whitespace before the
last 'v' line load incorrectly.

pinwheel.obj was changed to reflect this(!)